# ARROW-17824: [C++][Gandiva] Implement preallocation for variable length output buffer