@js8544 2022-09-24
I've tested the performance with a newly added benchmark TimedTestOutputStringAllocs. 
On my machine, it changes from 361038us to 290553us for jemalloc, and 6765325us to 291024us. Huge improvement for mimalloc.